Tblog/requirement.txt
aliyunsdkcore==1.0.3
amqp==2.5.2
anyjson==0.3.3
billiard==3.6.3.0
celery==4.4.2
celery-with-redis==3.0
Django==2.2
django-celery-results==1.2.1
django-ranged-response==0.2.0
django-redis==4.11.0
django-simple-captcha==0.5.12
django-tinymce==2.8.0
djangorestframework==3.11.0
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.0.0
pycrypto==2.6.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
